+++ aics-domain/src/main/java/kgu/developers/domain/file/infrastructure/repository/QueryFileRepository.java
@@ -1,4 +1,4 @@
-package kgu.developers.domain.file.infrastructure;
+package kgu.developers.domain.file.infrastructure.repository;
 
 import static kgu.developers.domain.file.infrastructure.entity.QFileJpaEntity.fileJpaEntity;
